Extend python bindings and add factorized groups.
Extends the python bindings to include a REDUNDANT_OKAY flag.

Changes group behavior to allow factorization; the group passed
into the solver now includes all group ids that are factors. This
requires group ids to be non-zero. Python bindings changed so the
default group id is 2 instead of 0 to avoid divide-by-zero errors.