Added Min and Max to Expression language
Added MIN and MAX operations to the expression language. This allows for
bounding boxes and other similar geometric relationships to be
expressed. The current impelementation prioritizes the first argument in
partial derivative calculations in case of equality.

Tested using a system of 3 points constrained to be an equilateral
triangles with sides of length 1 and axis-aligned bounding box in the
third quadrant.

Found that the solver optimization of trying to solve equations "alone"
needed to be disabled for the test case to converge (and it converges to
"REDUNDANT_OKAY", so this optimization has been commented out.