Initial external C++ expression API implementation
Implemented expression trees in the external C++ API.

Ported the equilateral triangle in quadrant III example. This version
differs from the previous test in that the points are true entities,
there is a 3D (but fixed) workplane, and the point-distance constraints
are implemented using constraint objects instead of equations.

Currently not converging after 50 iterations, however the constraint
equations should be the same as before (barring the exact definition of
point-point distance) since the bounding box constraints are expressed
in 2D rather than 3D.