Added debugging tools.
Added #define controllable debugging print statements to help diagnose
problems while running in python. Defining DEBUG_CPP will cause debug
messages to be printed to stderr so they are visible in the Jupyter
kernel's terminal output. This is currently set to true in lib.cpp.

Also changed setup.py so that it always copies and re-builds C++ sources
to make it easier to rebuild while debugging since the python API is
fairly static at this point.