Added name passing into equational constraints.
Since equational constraints are added by the Expression and not by the
system (to get around gil limitations), they could not pass a name
string into the add_constraint function of the Python wrapper. This
commit adds an is_wrapper function to SolverSystem's in order to
distinguish the AIDL wrapper from the Cython one (True for AIDL, False
for Cython), and will call the extended version of add_constraint if the
AIDL one is being used.